requirements.txt
yacs
pytorch-ignite==0.1.2
tqdm==4.54
scikit-learn==1.3.2
numpy==1.24.4
timm==0.3.2
ftfy
regex